OK, RDDShare System Version 2.0: It can run in cluster mode.

+++ main/scala/rddShare/core/CacheMetaData.scala
@@ -12,9 +12,9 @@ class CacheMetaData(
       val nodesList: Array[SimulateRDD],            // DAG图的各个节点
       val indexOfDagScan: util.ArrayList[Integer],  // the leaf nodes(read file) of this DAG
       val outputFilename: String,                   // 结果保存的文件名
-      val outputFileLastModifiedTime: Double,       // use to maintain consistency
+      val outputFileLastModifiedTime: Long,       // use to maintain consistency
       val sizeOfOutputData: Double,
-      val exeTimeOfDag: Double,
+      val exeTimeOfDag: Long,
       val insertTime: Timestamp
      ) extends Serializable {
